pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "seasonality-analyzer"
version = "0.1.0"
description = "CLI tool for detecting seasonality patterns in time series data"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Seasonality Analyzer Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Information Analysis",
]
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "statsmodels>=0.14.0",
    "scikit-learn>=1.3.0",
    "ruptures>=1.1.8",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "click>=8.1.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "PyYAML>=6.0.0",
    "loguru>=0.7.0",
    "tqdm>=4.65.0",
    "astropy>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]
llm = [
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]
full = [
    "prophet>=1.1.4",
    "plotly>=5.15.0",
    "jupyter>=1.0.0",
    "ipywidgets>=8.0.0",
    "openpyxl>=3.1.0",
    "psutil>=5.9.0",
]

[project.scripts]
seasonality = "seasonality.cli:main"

[project.urls]
Homepage = "https://github.com/example/seasonality-analyzer"
Repository = "https://github.com/example/seasonality-analyzer"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::FutureWarning",
]

[tool.coverage.run]
source = ["src/seasonality"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
